Remove i64x2 <-> f64x2 conversion opcodes
In the SIMD sync on 2020-01-14, we discussed the challenges with
implementing these instructions in a performant way on 32-bit archs. And
even on 64-bit arch, like x64, it requires AVX-512 for good codegen.
There is consensus that because it is uncommon for such instructions to
be used, and that hardware support is not widespread, we can remove
these.