Support i16x8 and implement neg, add, sub, mul
Create a new module I16, which uses Int.Make, and is backed by Int32. It
reuses a bunch of logic in Int. It stores 16-bit integers sign-extended
in Int32. This means that -1 (0xFFFF) is stored as 0xFFFFFFFF, rather
than 0x0000FFFF. All the bytes decode/encode logic is also done using
the signed form.