Set OCaml version to 4.08.1
The Bytes [0] functions we use require 4.08.1. This should only be set
on SIMD repo, since bytes is only used for SIMD implementation. By the
time we would like to merge this back into main spec, 4.08.1 should
hopefully be widespread enough for us to bump minimum version of OCaml
for spec to 4.08.1 too.

[0] https://caml.inria.fr/pub/docs/manual-ocaml/libref/Bytes.html see
"Binary encoding/decoding of integers"