Implement v8x16.swizzle
Decided to cheat a little bit, and squeeze v8x16.swizzle into the i8x16
ast node, so that we don't have to create a new v8x16 data type.
Semantically it means the same thing, v8x16 is called that way since it
doesn't treat the underlying data as any particular type, so treating it
as int will work too.