Implement SIMD load splat and load extend
Introduce a new SimdLoad AST node to represent SIMD load splat and load
extend. The existing parameters for loadop is not sufficient to
represent all the load splats and extends, as it only has 3 pack sizes,
and 2 sign-extensions, and 1 none case, giving at total of 7 cases:

- sz = (pack_size * extension) option

while we have 6 load extends and 4 load splats (10 in total).

We can add 1 more case to pack_size, Pack64, which will only be used by
SIMD loads, but that still only gives 8 cases. We can also add 1 more
data type to extension, that will give us enough, but it will be useless
for non-v128 loads.