Run simd test cases with make partest
We list all simd test cases in simd/* in TESTFILES, then tweak the
substitution in quiettest to strip quiettest/ from the front of the
target, rather than match on just the file name (which loses the simd/)
prefix.

The temp output file name is unchanged, it uses the base file name (no
simd/ prefix).